requirements.txt
websockets>=12.0
asyncpg>=0.29.0
httpx>=0.27.0
sentence-transformers>=3.0.0
ddgs>=9.0.0
docker>=7.0.0
paddleocr>=2.7.0
paddlepaddle-gpu>=2.6.0
ultralytics>=8.2.0
transformers>=4.40.0
Pillow>=10.0.0
numpy>=1.24.0
